Add turning tank
Now if pres wasd, tank can turn. I remove 4 "put" methods with methods
that change corrent turn position _curPos by passing a reference to an
array of the corresponding rotation position _up/down/left/rightPos.
Now, for me, this is the simplies was the most obvious way realization,
so I did it in this way, maybe I will change it later, if I find better way.