First version of Player class
This class inherited from Tank class, I moved the logic of tank control
from index.ts to this class, all other are from the base class. In
future will be the Enemy class, so It's better to separate controls from
basic Tank class, because enemy will be controled by some algorithm or
something like this.